requirements.txt
Python 3.8
torch>=2.0.0
torchvision>=0.15.0
numpy
Pillow
tqdm
opencv-python
matplotlib
